requirements.txt
streamlit>=1.32.0
openai>=1.12.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pytest>=8.0.0
python-dateutil>=2.9.0
pyyaml>=6.0
requests>=2.31.0
